aula 05
Implementando swagger

+++ src/main/java/br/com/fiap/api_rest/controller/ClienteController.java
@@ -18,7 +18,8 @@ import java.util.Optional;
 import java.util.function.Function;
 
 @RestController
-@RequestMapping("/clientes")
+@RequestMapping(value = "/clientes", produces = {"application/json")
+@Tag(name = "api-clientes")
 public class ClienteController {
     @Autowired
     ClienteRepository clienteRepository;
@@ -27,12 +28,21 @@ public class ClienteController {
 
     // CREATE READ UPDATE DELETE
     // POST GET PUT DELETE
+    @Operation(summary = "Cria um novo cliente")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "Cliente cadastrado com sucesso",
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = Cliente.class))}),
+            @ApiResponse(responseCode = "400", description = "Atributos informados são inválidos",
+                    content = @Content(schema = @Schema()))
+    })
     @PostMapping
     public ResponseEntity<Cliente> createCliente(@Valid @RequestBody ClienteRequest cliente){
         Cliente clienteSalvo = clienteRepository.save(clienteService.requestToCliente(cliente));
         return new ResponseEntity<>(clienteSalvo, HttpStatus.CREATED);
     };
 
+    @Operation(summary = "Retorna uma lista de clientes")
     @GetMapping
     public ResponseEntity<Page<ClienteResponse>> readClientes(@RequestParam(required = true) int page) {
         Pageable pageable = PageRequest.of(page, 2, Sort.by("categoria").ascending().and(Sort.by("nome").ascending()));
@@ -41,6 +51,15 @@ public class ClienteController {
 
     // PathVariable = parâmetro URL EX: /clientes/1
     // RequestParam = parâmetro query EX: /clientes/?id=1
+    @Operation(summary = "Retorna um cliente por ID")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Cliente encontrado com sucesso",
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = ClienteResponse.class))}),
+            @ApiResponse(responseCode = "404", description = "Nenhum cliente encontrado",
+                    content = @Content(schema = @Schema()))
+    })
+
     @GetMapping("/{id}")
     public ResponseEntity<ClienteResponse> readCliente(@PathVariable Long id) {
         Optional<Cliente> cliente = clienteRepository.findById(id);
@@ -50,6 +69,14 @@ public class ClienteController {
         return new ResponseEntity<>(clienteService.clienteToResponse(cliente.get(), true), HttpStatus.OK);
     }
 
+    @Operation(summary = "Atualiza um cliente existente")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "Cliente encontrado e atualizado com sucesso",
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = Cliente.class))}),
+            @ApiResponse(responseCode = "400", description = "Nenhum cliente encontrado para atualizar",
+                    content = @Content(schema = @Schema()))
+    })
     @PutMapping("/{id}")
     public ResponseEntity<Cliente> updateCliente(@PathVariable Long id,
                                                  @RequestBody Cliente cliente) {
@@ -61,9 +88,19 @@ public class ClienteController {
         Cliente clienteAtualizado = clienteRepository.save(cliente);
         return new ResponseEntity<>(clienteAtualizado, HttpStatus.CREATED);
     }
-
+    @Operation(summary = "Exclui um cliente por ID")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "400", description = "Nenhum cliente encontrado para excluir",
+                    content = @Content(schema = @Schema())),
+            @ApiResponse(responseCode = "204", description = "Cliente excluído com sucesso",
+                    content = @Content(schema = @Schema()))
+    })
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteCliente(@PathVariable Long id) {
+        Optional<Cliente> clienteExistente = clienteRepository.findById(id);
+        if (clienteExistente.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         clienteRepository.deleteById(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
